A rough parser for Rebellion's Asura File Format
main and check_bytes are both main files I was using to examine/parse the files.
This is not production ready, and can only read files.
HText files can be read, and should be able to be written. The primary issue would be the padding for the KEY before the internal name list. (may be next 4th byte? needs investigation)

=== === ===
Special thanks to the people who examined Sniper Elite & Alien (2010) on XeNTaX for their work on cracking earlier Asura files, it made this a lot easier once I found my footing.

Special thanks to texconv (DirectX Converter) for saving me from wasting more sleep on wondering why a TGA file is actully a DDS (DX10 format) file that I can't open, by letting me convert it to a tiff and knowing that my efforts were not wasted.