Reorg + Massive Fixes
Fixed HTEXT and Resource decoding;
As far as I can tell; reading works flawlessly for these files.

I've figured out one of the mystery words in HTEXT is language id (although a XeNTaX doc did mention it), confirmed through the use of steam's language settings

Models & Enums were moved into a package to make it easier to find; access doesn't change.

Began writing a reader/writer wrapper but I'm still not sure how I'd like to handle it.
In C# I'd just implement the interface; and proxy it (the IDE would do most of the work). But Python's proxy stuff is a bit more complicated; and I'd prefer not to use the __getattr_ method without understanding the consequences a bit more