The Price I Pay
This is a problem I would love to solve in C#; but I started this in python; I'll finish it in python.
~ ~ ~
I'll come out and say it; I don't like python's module setup (but i understand why), coming from C#, I like the 'one file one purpose' mentality. I want to do that in python, but some of my classes have tight coupling. In C#, that's fine and dandy; in python I get a mess of cyclic dependency errors.

So I need to get to the crux of the issue; decoupling.