I broke unpacker
Making an unpacker that's nice and that can be reversed is hard.

An easy solution is Archive (folder) / Chunk (folder/file) + meta / Chunk Format Files (most likely resources)

This works, but resources-group chunks are isolated to their chunks, making it hard to view all resources.

A nice solution (imo) would be:
Archive (folder)
- Chunks (folder/file) + meta
-- Chunk Format Files
- Resources (folder)
-- Resource format files

This makes it easy to find and edit files, but repacking archives is more difficult; adding files requires adding information in the chunks folder.