Working?
It requires a full archive dump, but the unpacker/repacker are capable of completely unpacking and then repacking the assets

With any luck, I can start modding assets without writing individual scripts to edit specific files. And with any luck, it will actually work.

While I understand the gist of the archive formats, I do not understand how asura handles resources, they appear to be 'pathed' which means the engine has to load them and cache them, which makes sense (Unity's asset bundles are the best analogy I can think of right now).

But manually editing uncompressed archives with audio doesn't do anything; the audio doesn't play. Whether this is because it isn't loaded (at all or properly) or because it isn't understood, I can't be certain.  If its the former; then I need to figure out the unknowns in their formats. If it's the latter,  then it may be a wave-format mismatch (which I doubt since all files are PCM 2, both custom PCM 1 and 2 failed, and I'd be baffled if the engine couldn't handle raw waves (unless it is because of size limits?))