Unpacking -> Editing -> Repacking Zbb Test
Image flipper tries to flip all images in an archive to see if any files use checksums.
This commit does not have a working version (due to recompression it seems)

Tests are nice when you want to run them; a commit following this should make it so tests only run on the test branch.