Uh Oh ~ Incompatable wwith AvP:2010
I figured it might be too easy to use the model format AvP:2010 used, but I was hopeful.

At the very least; (if this format is similar) I've learned how to expect mesh files to be written.

After testing with the official bms script and the viewer; it doesn't work so I'm pretty confident that the format's changed. (0x00 0x01 being the type might have something to do with it. Maybe the second byte is exclusively used for the mesh format?
Source for alot of this is from Trololp's AvP:2010 model viewer
https://github.com/Trololp/AVP2010ModelViewer